+++ xarray/core/duck_array_ops.py
@@ -631,3 +631,27 @@ def least_squares(lhs, rhs, rcond=None, skipna=False):
         return dask_array_ops.least_squares(lhs, rhs, rcond=rcond, skipna=skipna)
     else:
         return nputils.least_squares(lhs, rhs, rcond=rcond, skipna=skipna)
+
+
+def push(array, n, axis):
+    """
+    Dask-aware bottleneck.push
+    """
+    from bottleneck import push
+
+    if is_duck_dask_array(array):
+        if len(array.chunks[axis]) > 1 and n is not None and n < array.shape[axis]:
+            raise NotImplementedError(
+                "Cannot fill along a chunked axis when limit is not None."
+                "Either rechunk to a single chunk along this axis or call .compute() or .load() first."
+            )
+        if all(c == 1 for c in array.chunks[axis]):
+            array = array.rechunk({axis: 2})
+        pushed = array.map_blocks(push, axis=axis, n=n)
+        if len(array.chunks[axis]) > 1:
+            pushed = pushed.map_overlap(
+                push, axis=axis, n=n, depth={axis: (1, 0)}, boundary="none"
+            )
+        return pushed
+    else:
+        return push(array, n, axis)